+++ src/4 Static Equilibrium/p47.jl
@@ -17,8 +17,8 @@ p47(data)
 ### Required data dictionary keys
 ```julia
 * struc_el::StructuralElement                          : Structural element
-* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
-* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
+* support::Array{Tuple{Int,Array{Int,1}},1}        : Fixed-displacements vector
+* loaded_nodes::Array{Tuple{Int,Array{Float64,1}},1} : Node load vector
 * properties::Vector{Float64}                          : Material properties
 * x_coords::FloatRange{Floalt64}                       : x-coordinate vector
 * y_coords::FloatRange{Floalt64}                       : y-coordinate vector
@@ -28,7 +28,7 @@ p47(data)
 ### Optional additional data dictionary keys
 ```julia
 * penalty = 1e20               : Penalty used for fixed degrees of freedoms
-* etype::Vector{Int64}         : Element material vector if np_types > 1
+* etype::Vector{Int}         : Element material vector if np_types > 1
 ```
 
 ### Return values
@@ -105,7 +105,7 @@ function p47(data::Dict{Symbol, Any})
     println("No :properties key found in FEdict")
   end
     
-  nf = ones(Int64, nodof, nn)
+  nf = ones(Int, nodof, nn)
   if :support in keys(data)
     for i in 1:size(data[:support], 1)
       nf[:, data[:support][i][1]] = data[:support][i][2]
@@ -127,7 +127,7 @@ function p47(data::Dict{Symbol, Any})
     z_coords = data[:z_coords]
   end
 
-  etype = ones(Int64, nels)
+  etype = ones(Int, nels)
   if :etype in keys(data)
     etype = data[:etype]
   end
@@ -135,13 +135,13 @@ function p47(data::Dict{Symbol, Any})
   # All other arrays
   
   points = zeros(struc_el.nip, ndim)
-  g = zeros(Int64, ndof)
+  g = zeros(Int, ndof)
   g_coord = zeros(ndim,nn)
   fun = zeros(fin_el.nod)
   coord = zeros(fin_el.nod, ndim)
   gamma = zeros(nels)
   jac = zeros(ndim, ndim)
-  g_num = zeros(Int64, fin_el.nod, nels)
+  g_num = zeros(Int, fin_el.nod, nels)
   der = zeros(ndim, fin_el.nod)
   deriv = zeros(ndim, fin_el.nod)
   bee = zeros(nst,ndof)
@@ -151,8 +151,8 @@ function p47(data::Dict{Symbol, Any})
   kg = zeros(ndof, ndof)
   eld = zeros(ndof)
   weights = zeros(struc_el.nip)
-  g_g = zeros(Int64, ndof, nels)
-  num = zeros(Int64, fin_el.nod)
+  g_g = zeros(Int, ndof, nels)
+  num = zeros(Int, fin_el.nod)
   actions = zeros(ndof, nels)
   displacements = zeros(size(nf, 1), ndim)
   gc = ones(ndim)
@@ -162,7 +162,7 @@ function p47(data::Dict{Symbol, Any})
   
   formnf!(nodof, nn, nf)
   neq = maximum(nf)
-  kdiag = zeros(Int64, neq)
+  kdiag = zeros(Int, neq)
   
   # Find global array sizes
   
@@ -226,9 +226,9 @@ function p47(data::Dict{Symbol, Any})
   if :fixed_freedoms in keys(data)
     fixed_freedoms = size(data[:fixed_freedoms], 1)
   end
-  no = zeros(Int64, fixed_freedoms)
-  node = zeros(Int64, fixed_freedoms)
-  sense = zeros(Int64, fixed_freedoms)
+  no = zeros(Int, fixed_freedoms)
+  node = zeros(Int, fixed_freedoms)
+  sense = zeros(Int, fixed_freedoms)
   value = zeros(Float64, fixed_freedoms)
   if :fixed_freedoms in keys(data) && fixed_freedoms > 0
     for i in 1:fixed_freedoms
